refactor(CalculatorType): Change format of CalculatorType Enum
