+++ src/main/scala/com/particeep/api/models/enums/CalculatorType.scala
@@ -1,13 +1,11 @@
 package com.particeep.api.models.enums
 
-object CalculatorType {
+sealed trait CalculatorType extends Product with Enum
 
-  sealed abstract class CalculatorType extends Product with Enum
-  case object Constant                 extends CalculatorType
-  case object CouponConstant           extends CalculatorType
-  case object InFine                   extends CalculatorType
+object CalculatorType extends EnumHelper[CalculatorType] {
+  case object Constant       extends CalculatorType
+  case object CouponConstant extends CalculatorType
+  case object InFine         extends CalculatorType
 
-  object CalculatorType extends EnumHelper[CalculatorType] {
-    def values: Set[CalculatorType] = Set(Constant, CouponConstant, InFine)
-  }
+  def values: Set[CalculatorType] = Set(Constant, CouponConstant, InFine)
 }
