+++ src/main/scala/com/particeep/api/models/fundraise/loan/RepaymentInfo.scala
@@ -5,7 +5,7 @@ import play.api.libs.json.{ Json, OFormat, Writes }
 import java.time.OffsetDateTime
 
 import com.particeep.api.core.Formatter
-import com.particeep.api.models.enums.CalculatorType.CalculatorType
+import com.particeep.api.models.enums.CalculatorType
 
 case class RepaymentInfo(
   method:               CalculatorType,
